dzil-actions/build: no more v5.8 testing
Test2::Harness and the JUnit formatter now require v5.10.0 and v5.10.1
respectively.  Code being tested might work on v5.8.0, but I am not
interested in changing the test infrastructure for the sake of such an
obsolete perl.  With this change, we might still ship something that can
run on v5.8, but it won't be tested (or supported).